data_ingester/.deprecated/data_ingester/requirements.txt
boto3
requests
dnspython
msgraph-beta-sdk
microsoft-kiota-serialization-json
microsoft-kiota-abstractions
azure-identity
azure-keyvault-secrets
